backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pydantic
python-dotenv
google-generativeai
openai
python-multipart
duckduckgo-search
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
gunicorn
